+++ server/routes/mycelium.js
@@ -206,6 +206,7 @@ import { registerRequestRoutes } from './requests.js';
 import { registerEventRoutes } from './events.js';
 import { registerAgentRoutes } from './agents.js';
 import { registerInboxRoutes } from './inbox.js';
+import { registerProfileRoutes } from './profiles.js';
 
 var ADMIN_KEY = process.env.ADMIN_KEY;
 function isAdminKey(key) {
@@ -2967,77 +2968,11 @@ router.post('/github/prs/:owner/:repo', asyncHandler(function (req, res) {
     .catch(function (e) { console.error('[mycelium] GitHub API error:', e.message); res.status(500).json({ error: 'GitHub request failed' }); });
 }));
 
-// ======== NODE PROFILES (Stand Up Calibration) ========
+// ======== NODE PROFILES (extracted to profiles.js) ========
 
-// List all profiles (admin only)
-router.get('/profiles', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  var filter = {};
-  if (req.query.node_type) filter.node_type = req.query.node_type;
-  if (req.query.layer) filter.layer = req.query.layer;
-  var profiles = listNodeProfiles(filter);
-  res.json({ count: profiles.length, profiles: profiles });
-}));
-
-// Resolve profile chain for an agent (admin only)
-// NOTE: This route must be before /profiles/:id to avoid matching "resolve" as an ID
-router.get('/profiles/resolve/:agentId', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  var resolved = resolveProfileChain(req.params.agentId);
-  res.json(resolved);
-}));
-
-// Get single profile (admin only)
-router.get('/profiles/:id', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  var profile = getNodeProfile(req.params.id);
-  if (!profile) return res.status(404).json({ error: 'Profile not found' });
-  res.json(profile);
-}));
-
-// Create profile (admin only)
-router.post('/profiles', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  var id = req.body.id;
-  if (!id) return res.status(400).json({ error: 'id is required' });
-  // Check if profile already exists
-  var existing = getNodeProfile(id);
-  if (existing) return res.status(409).json({ error: 'Profile already exists: ' + id });
-  try {
-    var profile = createNodeProfile(id, req.body);
-    emitEvent('profile_created', getAdminDisplayName(req), null, 'Profile created: ' + id);
-    res.status(201).json(profile);
-  } catch (e) {
-    console.error('[mycelium] profile creation error:', e.message);
-    res.status(500).json({ error: 'Failed to create profile' });
-  }
-}));
-
-// Update profile (admin only, partial)
-router.put('/profiles/:id', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  var updated = updateNodeProfile(req.params.id, req.body);
-  if (!updated) {
-    var existing = getNodeProfile(req.params.id);
-    if (!existing) return res.status(404).json({ error: 'Profile not found' });
-    return res.status(403).json({ error: 'Cannot modify platform-layer profiles' });
-  }
-  emitEvent('profile_updated', getAdminDisplayName(req), null, 'Profile updated: ' + req.params.id);
-  res.json(updated);
-}));
-
-// Delete profile (admin only, blocked for platform layer)
-router.delete('/profiles/:id', asyncHandler(function (req, res) {
-  if (!checkAdmin(req, res)) return;
-  var deleted = deleteNodeProfile(req.params.id);
-  if (!deleted) {
-    var existing = getNodeProfile(req.params.id);
-    if (!existing) return res.status(404).json({ error: 'Profile not found' });
-    return res.status(403).json({ error: 'Cannot delete platform-layer profiles' });
-  }
-  emitEvent('profile_deleted', getAdminDisplayName(req), null, 'Profile deleted: ' + req.params.id);
-  res.json({ ok: true, deleted: deleted });
-}));
+registerProfileRoutes(router, {
+  asyncHandler, checkAdmin, emitEvent, getAdminDisplayName,
+});
 
 // ======== TEAM SETTINGS ========
